# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/resources": {
      "get": {
        "summary": "Get all resources",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of resources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                },
                "example": [
                  {
                    "name": "Gold",
                    "price": 100
                  },
                  {
                    "name": "Silver",
                    "price": 50
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new resource",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Resource"
              },
              "example": {
                "name": "Gold",
                "price": 100
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resource created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                },
                "example": {
                  "name": "Gold",
                  "price": 100
                }
              }
            }
          }
        }
      }
    },
    "/resources/{id}": {
      "get": {
        "summary": "Get a resource by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resource found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                },
                "example": {
                  "name": "Gold",
                  "price": 100
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a resource by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resource deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                },
                "example": {
                  "name": "Gold",
                  "price": 100
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "put": {
        "summary": "Update a resource by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Resource"
              },
              "example": {
                "name": "Gold",
                "price": 150
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resource updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                },
                "example": {
                  "name": "Gold",
                  "price": 150
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "patch": {
        "summary": "Update the price of a resource by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "price": {
                    "type": "number",
                    "example": 200
                  }
                },
                "required": ["price"]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resource price updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                },
                "example": {
                  "name": "Gold",
                  "price": 200
                }
              }
            }
          },
          "400": {
            "description": "Price is required"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/resources/{id}/resource": {
      "get": {
        "summary": "Get specific resource amount for authenticated user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resource amount for user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resourceId": {
                      "type": "string",
                      "example": "1"
                    },
                    "name": {
                      "type": "string",
                      "example": "Gold"
                    },
                    "price": {
                      "type": "number",
                      "example": 100
                    },
                    "amount": {
                      "type": "number",
                      "example": 10
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found for this user"
          }
        }
      },
      "post": {
        "summary": "Create new user-resource link",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "User-resource link created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resourceId": {
                      "type": "string",
                      "example": "1"
                    },
                    "name": {
                      "type": "string",
                      "example": "Gold"
                    },
                    "price": {
                      "type": "number",
                      "example": 100
                    },
                    "amount": {
                      "type": "number",
                      "example": 10
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Resource already linked to user or Resource not found"
          }
        }
      },
      "patch": {
        "summary": "Update resource amount for authenticated user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "example": 20
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resource amount updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resourceId": {
                      "type": "string",
                      "example": "1"
                    },
                    "name": {
                      "type": "string",
                      "example": "Gold"
                    },
                    "price": {
                      "type": "number",
                      "example": 100
                    },
                    "amount": {
                      "type": "number",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource link not found"
          }
        }
      }
    },
...
    "/resources/stats": {
      "get": {
        "summary": "Get resource statistics per user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resource statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResourceStats"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/resources/init": {
      "post": {
        "summary": "Initialize all resources for authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Resources initialized successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "resources": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserResource"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/resources/user/resources": {
      "get": {
        "summary": "Get all resources with amounts for authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User resources retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserResource"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Resource": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Gold"
          },
          "price": {
            "type": "number",
            "example": 100
          }
        },
        "required": ["name", "price"]
      },
...
      "UserResource": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          }
        }
      },
      "UserResourceStats": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "totalResources": {
            "type": "number"
          },
          "totalValue": {
            "type": "number"
          },
          "resourceCount": {
            "type": "number"
          },
          "resources": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "value": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}